// The-Food-Works-Mobile/src/app/interfaces/product.ts
export interface IProductMenuItem {
  productID: string;
  productImage: string;
  productName: string;
  productRating: number;
  productPrice: number;
}

export interface IProduct {
  productID: number;
  productImage: string;
  productName: string;
  productDescription: string;
  productRating: number;
  productPrice: number;
}